pyproject.toml
[tool.poetry]
name = "jumptofastapi"
version = "0.1.0"
description = ""
authors = ["Kim Sanghae <pelikan@pusan.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-dateutil = "^2.9.0.post0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
